Fix not reading partial responses correctly
When dealing with a separate header and payload, make sure that we
do not expect to receive additional USB packet headers in the payload portion of
the transfer. Previously, we were assuming that this would only occur when the
device was sending more than a full packet's worth of data, but this is not correct -
we should continue the bulk transfer any time the device sends us less data than
the header says it will be sending.